+++ macros/tests/ui/enum_not_supported.rs
@@ -1,5 +1,5 @@
 // This test should fail to compile because enums are not supported
-use stardrift_macros::ConfigDefaults;
+use macros::ConfigDefaults;
 
 #[derive(ConfigDefaults)]
 enum NotSupported {