+++ macros/tests/ui/tuple_struct.rs
@@ -1,5 +1,5 @@
 // This test should fail to compile because tuple structs are not supported
-use stardrift_macros::ConfigDefaults;
+use macros::ConfigDefaults;
 
 #[derive(ConfigDefaults)]
 struct TupleStruct(i32, String);