+++ macros/tests/ui/unit_struct.rs
@@ -1,5 +1,5 @@
 // This test should fail to compile because unit structs are not supported
-use stardrift_macros::ConfigDefaults;
+use macros::ConfigDefaults;
 
 #[derive(ConfigDefaults)]
 struct UnitStruct;